refactor: extract segment copy job builder into transfer-segment-copy module
